Refactor FireCoverageRenderer to FireCoverageLayer
