+++ src/main/kotlin/kotcity/ui/QueryWindow.kt
@@ -10,6 +10,7 @@ import kotcity.data.CityMap
 import kotcity.data.Road
 import kotcity.ui.sprites.BuildingSpriteLoader
 import tornadofx.View
+import kotlin.math.roundToInt
 
 class QueryWindow : View() {
     override val root: BorderPane by fxml("/QueryWindow.fxml")
@@ -70,7 +71,7 @@ class QueryWindow : View() {
                 buffer.append("Elevation: ${city.groundLayer[coordinate]?.elevation}\n")
                 buffer.append("Land Value: ???\n")
 
-                buffer.append("Fire Coverage: ${city.fireCoverageLayer.readableValue(coordinate)}\n")
+                buffer.append("Fire Coverage: ${(city.fireCoverageLayer[coordinate] ?: 0.0).roundToInt() * 100} %\n")
 
                 buildingDescriptionArea.text = buffer.toString()
             }
